+++ libs/sdk-core/src/lib/db/search/chat.ts
@@ -1,12 +1,12 @@
 import { catchError, map, Observable, of, switchMap } from 'rxjs';
 import type { IErrorResponse, INuclia } from '../../models';
-import { Chat } from './chat.models';
+import { Chat, Citations } from './chat.models';
 import { ChatOptions, Search } from './search.models';
 
 import { ResourceProperties } from '../db.models';
-import { find } from './search';
 
 const END_OF_SEARCH_RESULTS = '_END_';
+const START_OF_CITATIONS = '_CIT_';
 
 export function chat(
   nuclia: INuclia,
@@ -19,6 +19,7 @@ export function chat(
   let sourcesLength = 0;
   let sources: Search.FindResults | undefined;
   let relations: Search.Relations | undefined;
+  let citations: Citations | undefined;
   let text = '';
   const { synchronous, ...searchOptions } = options;
   const endpoint = `${path}/chat`;
@@ -31,7 +32,7 @@ export function chat(
   };
   return synchronous
     ? nuclia.rest
-        .post<{ answer: string; relations: Search.Relations; results: Search.FindResults }>(
+        .post<{ answer: string; relations: Search.Relations; results: Search.FindResults; citations: Citations }>(
           endpoint,
           body,
           undefined,
@@ -39,11 +40,12 @@ export function chat(
           true,
         )
         .pipe(
-          map(({ answer, relations, results }) => {
+          map(({ answer, relations, results, citations }) => {
             return {
               type: 'answer',
               text: answer,
               sources: { ...results, relations },
+              citations,
               incomplete: false,
               id: '',
             } as Chat.Answer;
@@ -56,8 +58,12 @@ export function chat(
           // /chat returns a readable stream structured as follows:
           // - 1st block: 4 first bytes indicates the size of the 2nd block
           // - 2nd block: a base64 encoded JSON containing the sources used to build the answer
-          // - 3rd block: the answer text, ended by "_END_"
-          // - 4th block (optionally): base64 encoded JSON containing the entities
+          // - 3rd block: the answer text, ended by "_CIT_" (or "_END_" if no citations)
+          // - 4th block (optionally):
+          //      - 4 first bytes indicates the size of the block,
+          //      - followed by a base64 encoded JSON containing the citations
+          //      - ended by "_END_"
+          // - 5th block (optionally): base64 encoded JSON containing the relations
           if (sourcesLength === 0 && data.length >= 4) {
             sourcesLength = new DataView(data.buffer.slice(0, 4)).getUint32(0);
           }
@@ -73,6 +79,16 @@ export function chat(
             if (text.includes(END_OF_SEARCH_RESULTS)) {
               let relationsBase64;
               [text, relationsBase64] = text.split(END_OF_SEARCH_RESULTS);
+              if (text.includes(START_OF_CITATIONS)) {
+                let citationsBlock;
+                [text, citationsBlock] = text.split(START_OF_CITATIONS);
+                const citationsBase64 = citationsBlock.slice(4); // remove the 4 first bytes indicating the size of the block
+                try {
+                  citations = JSON.parse(atob(citationsBase64));
+                } catch (e) {
+                  console.warn(e);
+                }
+              }
               if (relationsBase64) {
                 try {
                   relations = JSON.parse(atob(relationsBase64));
@@ -83,7 +99,7 @@ export function chat(
               }
             }
           }
-          return { type: 'answer', text, sources, incomplete, id } as Chat.Answer;
+          return { type: 'answer', text, sources, incomplete, id, citations } as Chat.Answer;
         }),
         catchError((error) =>
           of({ type: 'error', status: error.status, detail: error.detail || '' } as IErrorResponse),