+++ libs/sdk-core/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - `/ephemeral_tokens` endpoint can now be called using regional domain
 - Support `summarize` operation on resources
-- Remove `onlyAnswers` property from `WidgetFeatures` interface, only answers is now managed by `hideSources` option. 
+- Remove `onlyAnswers` property from `WidgetFeatures` interface, only answers is now managed by `hideSources` option.
+- Support the `citation` option on `/chat` endpoint
 
 ### Bugfix