+++ libs/sdk-core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.9.1 (unreleased)
+# 1.9.1 (2023-12-18)
 
 ### Improvements